requirements.txt
discord.py>=2.7.0
PyNaCl>=1.5.0
python-dotenv>=1.0.1
davey>=0.1.5